% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{add_failure}
\alias{add_failure}
\title{Add users to the failure log}
\usage{
add_failure(user_ids)
}
\arguments{
\item{user_ids}{One or most user IDS (\emph{not} screen names) to add
to the failure log.}
}
\description{
Also refreshes the failure log.
}
\keyword{internal}
