% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{add_request}
\alias{add_request}
\title{Add users to the request queue}
\usage{
add_request(user_ids)
}
\arguments{
\item{user_ids}{One or most user IDS (\emph{not} screen names) to add
to the failure log.}
}
\description{
Also refreshes the request queue.
}
\keyword{internal}
