% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{add_users_edge_data_to_cache}
\alias{add_users_edge_data_to_cache}
\title{Update information on a user already in the cache to include edge data}
\usage{
add_users_edge_data_to_cache(users)
}
\arguments{
\item{users}{}
}
\value{

}
\description{
Notes:
}
\details{
\itemize{
\item The calling function is responsible for making sure that the user
is in the cache, but only node data is present, no edge data
\item If user A and user B are friends, and we add both user A
and user B, we will duplicate any edges between A and B.
\item We only take the first 5,000 friends and 5,000 followers
for the time being. The idea is that for important edges,
the edge will get picked up from the other node. This may
miss out on edges between node both with huge follower
and following counts, but who cares about those.
}
}
