% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet.R
\name{cache_get_friends}
\alias{cache_get_friends}
\title{Title}
\usage{
cache_get_friends(
  users,
  n = 5000,
  retryonratelimit = FALSE,
  page = "-1",
  parse = TRUE,
  verbose = TRUE,
  token = NULL
)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
Title
}
