% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached-api-calls.R
\name{cache_lookup_users}
\alias{cache_lookup_users}
\title{Title}
\usage{
cache_lookup_users(users)
}
\arguments{
\item{users}{}
}
\value{

}
\description{
note: new users are gonna burn get_friends and get_followers tokens
so be careful here!
}
