% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{check_rate_limits}
\alias{check_rate_limits}
\title{Check rate limits for registered tokens}
\usage{
check_rate_limits()
}
\description{
There should be no need to do this, but it may make you feel better.
}
