% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clean_cache}
\alias{clean_cache}
\title{Remove suspended and deleted accounts from the cache}
\usage{
clean_cache()
}
\description{
\itemize{
\item Also deduplicate edges
}
}
\details{
Not yet implemented.
}
