% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Remove all sampled users from the cache}
\usage{
clear_cache()
}
\description{
Will ask you for confirmation before proceeding because
of the how time consuming it is to sample users.
}
\seealso{
Other cache management: \code{\link{have_neighborhood}}
}
\concept{cache management}
