% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{clear_failures}
\alias{clear_failures}
\title{Empty the failure log}
\usage{
clear_failures()
}
\description{
Empty the failure log
}
\seealso{
Other managing failures: \code{\link{get_failures}},
  \code{\link{rerequest_failures}}
}
\concept{managing failures}
