% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{clear_requests}
\alias{clear_requests}
\title{Empty the request queue}
\usage{
clear_requests()
}
\description{
Empty the request queue
}
\seealso{
Other managing requests: \code{\link{get_requests}},
  \code{\link{request}}
}
\concept{managing requests}
