% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{clear_tokens}
\alias{clear_tokens}
\title{Remove all registered tokens from the cache}
\usage{
clear_tokens()
}
\description{
Will ask you for confirmation before proceeding because
of the how irritating it is to register tokens in the
first place.
}
\seealso{
Other managing tokens: 
\code{\link{get_number_of_tokens}()},
\code{\link{register_token}()}
}
\concept{managing tokens}
