% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_cache}
\alias{export_cache}
\title{Export the cache}
\usage{
export_cache(path)
}
\arguments{
\item{path}{}
}
\description{
Saves the cache to a zipped file that can be imported
elsewhere. Exported cache includes both twitter data
and registered tokens. Import the cache on a different
computer with \code{\link[=import_cache]{import_cache()}}.
}
