% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{failure_path}
\alias{failure_path}
\title{Get the path to the failure log}
\usage{
failure_path()
}
\value{
Path to \code{.rds} file.
}
\description{
The failure log is a character vector of user IDs (never
screen names) that gets stored as an \code{.rds} file.
}
\keyword{internal}
