% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_edge_data}
\alias{get_edge_data}
\title{Export the edge data from the cache to an edge list}
\usage{
get_edge_data()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with two columns \code{to} and \code{from}.
Entries in each column are Twitter user IDs and have
type \code{\link[bit64:integer64]{bit64::integer64()}}. This breaks from the \code{rtweet}
tradition of storing user IDs as character vectors,
because character encoding user IDs quickly becomes
inefficient for even moderate sized caches, which can
contain millions of edges.

To my knowledge, you should be able to use these columns
just like you'd expect. If you are uncomfortable with
\code{\link[bit64:integer64]{bit64::integer64()}} columns, you can convert them to
character columns with \code{as.character()}. \strong{Do not}
convert the columns to a base integer or numeric vector
with \code{as.integer} or \code{as.numeric}: this will result in
data loss.
}
\description{
As your cache grows in size, this gets slow pretty
quickly. I recommend saving the results to an \code{.rds}
file for convenience. I also recommend subsetting
as much as possible to the part of the graph you
are interested in. \code{get_edge_data()} is typically
much slower than \code{\link[=get_node_data]{get_node_data()}}.
}
\details{
If you only need the graph and don't need any data
on the nodes, coerce immediately to an \code{igraph}
object with \code{igraph::from_edgelist()}.
}
\seealso{
Other exporting: \code{\link{get_neighborhood_as_tidygraph}},
  \code{\link{get_network_as_tidygraph}},
  \code{\link{get_node_data}}
}
\concept{exporting}
