% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{get_failure_ids}
\alias{get_failure_ids}
\title{Get user IDs from the failure log}
\usage{
get_failure_ids()
}
\value{
Character vector of users IDs that couldn't be
sampled.
}
\description{
Get user IDs from the failure log
}
\keyword{internal}
