% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{get_failures}
\alias{get_failures}
\title{Get screen names from the failure log}
\usage{
get_failures()
}
\value{
A character vector of screen names.
}
\description{
\strong{NOTE}: calls \code{rtweet::lookup_users()} without checking API
rate limits. Unless you have an enormous (90,000+) number of
failures, or have been using your tokens outside of \code{twittercache}
this is extremely unlikely to cause a problem.
}
\seealso{
Other managing failures: \code{\link{clear_failures}},
  \code{\link{rerequest_failures}}
}
\concept{managing failures}
