% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_neighborhood}
\alias{get_neighborhood}
\title{Get the user IDs of all nodes in a user's neighborhood}
\usage{
get_neighborhood(user)
}
\arguments{
\item{node}{Either the screen name of user ID of a Twitter user.}
}
\value{
A character vector of user IDs. May be empty if the user
doesn't follow anyone.
}
\description{
Get the user IDs of all nodes in a user's neighborhood
}
