% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_neighborhood_as_tidygraph}
\alias{get_neighborhood_as_tidygraph}
\title{Get the neighborhood around a seed as a tidygraph}
\usage{
get_neighborhood_as_tidygraph(seed, core_only = TRUE)
}
\arguments{
\item{seed}{Screen name of user to consider.}

\item{core_only}{Logical indicating whether or not to return
just the core nodes. A node in the "core" of the cache if you
explicitly requested it. Equivalently, if it has its own file
in the cache. Other nodes will appear in the graph because of
they are connected to nodes in the core. So we have complete
information about the friendships of nodes in the core, and
only very partial information about the remaining nodes.
Defaults to \code{TRUE}.}
}
\value{
The network as a \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}} object.
Does not contain information on protected users.
}
\description{
Takes a moment. We recommend saving the resulting graph in a \code{.rds}
file for fast access later on.
}
\seealso{
Other exporting: \code{\link{get_edge_data}},
  \code{\link{get_network_as_tidygraph}},
  \code{\link{get_node_data}}
}
\concept{exporting}
