% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{get_number_of_failures}
\alias{get_number_of_failures}
\title{Get the number of users in the failure log}
\usage{
get_number_of_failures()
}
\value{
An integer.
}
\description{
Get the number of users in the failure log
}
\seealso{
Other managing failures: \code{\link{clear_failures}},
  \code{\link{get_failures}},
  \code{\link{rerequest_failures}}
}
\concept{managing failures}
