% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_number_of_requests}
\alias{get_number_of_requests}
\title{Get the number of users in the request queue}
\usage{
get_number_of_requests()
}
\value{
An integer.
}
\description{
Get the number of users in the request queue
}
\seealso{
Other managing requests: 
\code{\link{clear_requests}()},
\code{\link{get_requests}()},
\code{\link{request}()}
}
\concept{managing requests}
