% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_number_of_tokens}
\alias{get_number_of_tokens}
\title{Check how many tokens are in your cache}
\usage{
get_number_of_tokens()
}
\value{
An integer.
}
\description{
Check how many tokens are in your cache
}
\seealso{
Other managing tokens: \code{\link{clear_tokens}},
  \code{\link{register_token}}
}
\concept{managing tokens}
