% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_registered_token}
\alias{get_registered_token}
\title{Get a specific token in the cache}
\usage{
get_registered_token(index)
}
\value{
A Twitter auth token.
}
\description{
Errors if there are no tokens, or the requested
token index is out of bounds.
}
\keyword{internal}
