% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_request_ids}
\alias{get_request_ids}
\title{Get user IDs from the request queue}
\usage{
get_request_ids()
}
\value{
Character vector of users IDs that we need
to sample.
}
\description{
Get user IDs from the request queue
}
\keyword{internal}
