% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_requests}
\alias{get_requests}
\title{Get screen names from the request queue}
\usage{
get_requests()
}
\value{
A character vector of screen names.
}
\description{
\strong{NOTE}: calls \code{rtweet::lookup_users()} without checking API
rate limits. Unless you have an enormous (90,000+) number of
requests, or have been using your tokens outside of \code{twittercache}
this is extremely unlikely to cause a problem.
}
\seealso{
Other managing requests: 
\code{\link{clear_requests}()},
\code{\link{get_number_of_requests}()},
\code{\link{request}()}
}
\concept{managing requests}
