% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_token_dir}
\alias{get_token_dir}
\title{Get the path to the token directory}
\usage{
get_token_dir()
}
\value{
Path to the token directory.
}
\description{
Each token is saved in the token directory as:
}
\details{
\itemize{
\item \verb{1.rds}
\item \verb{2.rds}
\item ...
}
}
\keyword{internal}
