% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{have_neighborhood}
\alias{have_neighborhood}
\title{Check if a user's entire neighborhood is in the cache}
\usage{
have_neighborhood(user)
}
\arguments{
\item{node}{Either the screen name of user ID of a Twitter user.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Check if a user's entire neighborhood is in the cache
}
\seealso{
Other cache management: \code{\link{clear_cache}}
}
\concept{cache management}
