% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{import_cache}
\alias{import_cache}
\title{Import the cache}
\usage{
import_cache(path, overwrite = FALSE)
}
\arguments{
\item{path}{Path to a zipped folder created by \code{\link[=export_cache]{export_cache()}}.}

\item{overwrite}{Logical indicating whether or not to overwrite
user data when already present in the cache. When \code{TRUE},
data in the cache file will overwrite data in the local
cache. Defaults to \code{FALSE}.}
}
\description{
Adds users and token from the zipped cache file to the
local cache.
}
