% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{in_cache}
\alias{in_cache}
\title{Check if a node is in the cache}
\usage{
in_cache(nodes)
}
\arguments{
\item{nodes}{User ids of Twitter users. \emph{Not} the screen name.}
}
\value{
A logical vector.
}
\description{
Check if a node is in the cache
}
\keyword{internal}
