% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{neighborhood_in_cache}
\alias{neighborhood_in_cache}
\title{Check if a user's entire neighborhood is in the cache}
\usage{
neighborhood_in_cache(user)
}
\arguments{
\item{node}{Either the screen name of user ID of a Twitter user.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Check if a user's entire neighborhood is in the cache
}
