% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{refresh_failures}
\alias{refresh_failures}
\title{Update the failure log}
\usage{
refresh_failures()
}
\description{
Removes duplicates and users that have been successfully
sampled from the failure log.
}
\keyword{internal}
