% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{refresh_requests}
\alias{refresh_requests}
\title{Refresh the request queue}
\usage{
refresh_requests()
}
\description{
Removes duplicates, users that we have failed to sample,
and users already in the cache.
}
\details{
A fundamental assumption is that a node cannot be
in both the request queue and the failure log at the same time!
}
\keyword{internal}
