% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{register_token}
\alias{register_token}
\title{Register a Twitter authentication token with twittercache}
\usage{
register_token(token)
}
\arguments{
\item{token}{A Twitter auth token token created by
\code{\link[rtweet:create_token]{rtweet::create_token()}}.}
}
\description{
If you try to register a token already in the cache,
you'll get a warning. That is, you are protected from
creating duplicate entries in the token cache.
}
\details{
For more details on how to get tokens please read
\code{vignette("auth", package = "rtweet")}.
}
\examples{

\dontrun{

# if you haven't registered your token with
# rtweet yet, do this

token <- rtweet::create_token(
  app = "FILL_THIS_IN",
  consumer_key = "FILL_THIS_IN",
  consumer_secret = "FILL_THIS_IN",
  access_token = "FILL_THIS_IN",
  access_secret = "FILL_THIS_IN"
)

# HOWEVER, you can register the same token
# with rtweet multiple times by accident
# if you are not careful

# to register tokens with twittercache that
# you have already registered with rtweet

rtweet_tokens <- rtweet::get_tokens()

for (token in rtweet_tokens)
  register_token(token)

}

}
\seealso{
Other managing tokens: \code{\link{clear_tokens}},
  \code{\link{get_number_of_tokens}}
}
\concept{managing tokens}
