% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Slate users for addition to the cache}
\usage{
request(screen_names, neighborhood = FALSE)
}
\arguments{
\item{screen_names}{A (character?) vector of screen names or user ids
to slate for addition to the cache.}

\item{neighborhood}{Logical indicating whether the neighbors of each
user should be added as well. Neighbors include both friends and
followers. Defaults to \code{FALSE}.}
}
\value{
Nothing, called for side effects
}
\description{
Slate users for addition to the cache
}
\examples{

\dontrun{

users <- c("blahasjaf80920930", "tim", "alexpghayes", "alexgphayes")
request(users, neighborhood = TRUE)
get_current_requests()

}

}
