% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request_path}
\alias{request_path}
\title{Get the path to the request queue}
\usage{
request_path()
}
\value{
Path to \code{.rds} file.
}
\description{
The request queue is a character vector of user IDs (never
screen names) that gets stored as an \code{.rds} file.
}
\keyword{internal}
