% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{rerequest_failures}
\alias{rerequest_failures}
\title{Re-request users in the failure log}
\usage{
rerequest_failures()
}
\description{
Also empties the failure log.
}
\seealso{
Other managing failures: 
\code{\link{clear_failures}()},
\code{\link{get_failures}()}
}
\concept{managing failures}
