% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{safe_get_friends}
\alias{safe_get_friends}
\title{Title}
\usage{
safe_get_friends(node, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\seealso{
Other API call: \code{\link{check_rate_limits}},
  \code{\link{find_token}},
  \code{\link{safe_get_followers}},
  \code{\link{safe_lookup_user}}
}
\concept{API call}
