% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{sample_twitter_graph}
\alias{sample_twitter_graph}
\title{Robustly sample from the twitter graph}
\usage{
sample_twitter_graph()
}
\description{
Once you have added users to the request queue, run
\code{sample_twitter_graph()}. That's it. Data about each
user is saved into the cache as soon it is received.
}
\details{
If your R session crashes, you lose internet
connection, whatever, it doesn't matter. Just run
\code{sample_twitter_graph()} again and the sampling will pick
up right where it left off.

Sampling will automatically use all registered tokens,
and will respect rates limits. When we reach an API
rate limit, we simply wait for the shortest amount of
time until the rate limit resets for the next
available token.
}
\seealso{
\code{\link[=register_token]{register_token()}}, \code{\link[=request]{request()}}
}
