% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twittergraph-package.R
\docType{package}
\name{twittercache-package}
\alias{twittercache}
\alias{twittercache-package}
\title{twittercache: Sample the Twitter Graph and Cache It Locally}
\description{
Facilitates robust sampling of the Twitter graph by saving
  data about observed nodes into a data cache. Requests for new nodes are
  smart and do not waste API requests on nodes already in the cache. The
  cache can then be exported to a number of data structures more
  appropriate for network analysis. Automatically manages multiple tokens
  to speed up sampling with staying with API rate limits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/twittercache}
  \item Report bugs at \url{https://github.com/alexpghayes/twittercache/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

}
\keyword{internal}
