% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{update_cache}
\alias{update_cache}
\title{Update a version 0.1.0 cache to a version 0.2.0 cache}
\usage{
update_cache()
}
\description{
Slow. Sets \code{sampled_at} column in node data based on the
most recent time the node data was modified. This
should be the time of sampling unless you've been
manually moving files around or something.
}
\details{
If this crashes halfway through (which should only
happen due to user interrupt, fingers crossed)
you \emph{should} be able to just run it again.

You only need to update your cache if you sampled
users using \code{twittercache} version 0.1.0.
}
