% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{user_in_cache}
\alias{user_in_cache}
\title{Check if a user is in the cache}
\usage{
user_in_cache(user)
}
\arguments{
\item{node}{Either the screen name of user ID of a Twitter user.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
TODO: Not vectorized (yet)
}
