% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{users_in_cache}
\alias{users_in_cache}
\title{Check if a user is in the cache}
\usage{
users_in_cache(users)
}
\arguments{
\item{users}{A vector screen names or user IDs of Twitter users.
All one or the other -- don't mix screen names and user IDs.}
}
\value{
A logical vector.
}
\description{
Check if a user is in the cache
}
\seealso{
Other cache management: 
\code{\link{delete_cache}()}
}
\concept{cache management}
