% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{visit}
\alias{visit}
\title{Visit a node in the twitter graph and add it to the cache}
\usage{
visit(node, force = FALSE, token = NULL)
}
\arguments{
\item{node}{The user id of a Twitter user. \emph{Not} the screen name.}

\item{force}{Logical indicating if the data for \code{node} should be
updated even if \code{node} is already in the cache. Defaults to \code{FALSE}.}
}
\value{
Nothing, called for side effects.
}
\description{
Visit a node in the twitter graph and add it to the cache
}
