% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_network_as_tidygraph}
\alias{get_network_as_tidygraph}
\title{Get the cached graph as tidygraph}
\usage{
get_network_as_tidygraph(core_only = FALSE)
}
\arguments{
\item{neighbors_only}{Logical indicating whether nodes TODO}
}
\value{
The network as a \code{\link[=tbl_graph]{tbl_graph()}} object. Removes protected nodes.
}
\description{
Takes a moment. We recommend saving the resulting graph in a \code{.rds}
file for fast access later on.
}
