% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{in_cache}
\alias{in_cache}
\title{Check if a node is in the cache}
\usage{
in_cache(node)
}
\arguments{
\item{node}{The user id of a Twitter user. \emph{Not} the screen name.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Check if a node is in the cache
}
