% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Add users to the request queue}
\usage{
request(screen_names, neighborhood = FALSE)
}
\arguments{
\item{screen_names}{A characters vector of screen names or user ids
to the to the request queue.}

\item{neighborhood}{Logical indicating whether the neighbors (both
friends and followers) of the users in \code{screen_names} should be
added the request queue as well. Defaults to \code{FALSE}.}
}
\description{
Note that this only adds users to the request queue, but
does not do any sampling of the Twitter graph itself. To
sample the Twitter graph, run \code{\link[=sample_twitter_graph]{sample_twitter_graph()}}. As
of version \code{0.2.0} we no longer add protected users to the
the request queue.
}
\seealso{
\code{\link[=sample_twitter_graph]{sample_twitter_graph()}}

Other managing requests: \code{\link{clear_requests}},
  \code{\link{get_number_of_requests}},
  \code{\link{get_requests}}
}
\concept{managing requests}
