% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{user_in_cache}
\alias{user_in_cache}
\title{Check if a user is in the cache}
\usage{
user_in_cache(users)
}
\arguments{
\item{users}{A vector screen names or user IDs of Twitter users.}
}
\value{
A logical vector.
}
\description{
Check if a user is in the cache
}
\seealso{
Other cache management: \code{\link{clear_cache}}
}
\concept{cache management}
